Fri, Mar 11, 2022, 3:10 PM +00:00

+++ finance/app.py
@@ -67,7 +67,7 @@ def buy():
         stock = lookup(request.form.get("symbol"))
         if stock:
             price = float(stock["price"])
-            quantity = float(request.form.get("shares"))
+            quantity = int(request.form.get("shares"))
             cost = price * quantity
             print(balance)
             # check balance
